# pos/openapi/RmsSupportHub.Pos.Agent.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "RMS+ POS Agent API",
    "description": "Destination-owned foundation contract for the per-device Windows POS Agent. The Agent exposes a fixed HTTPS loopback endpoint for direct Support Hub browser access. Windows Negotiate supplies the authenticated identity, local Built-in Administrators membership authorizes protected reads and typed Windows-service and RMS database operations, and no Support Hub backend relay is involved. The Agent exposes read-only device/configuration/RMS visibility plus allow-listed service control and server-owned RMS database backup/restore workflows. Downloader, maintenance, and artifact capabilities remain typed, principal-scoped, and fail-closed. Mutation tokens are short-lived, one-use, and bound to the authenticated principal, exact Origin, registered operation, target, method, and path.",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://rms-pos-agent.localhost:5001",
      "description": "Fixed HTTPS loopback Agent origin"
    }
  ],
  "paths": {
...
    "/api/v1/main-server/profiles": {
      "get": {
        "tags": [
          "Main Server"
        ],
        "summary": "Read fixed Main Server profiles",
        "description": "Returns only the server-owned Testing profile and disabled future Production profile. Branch/POS binding, credentials, free URLs, and arbitrary Main Server routes remain Agent-owned and are not exposed.",
        "operationId": "GetMainServerProfiles",
        "responses": {
          "200": {
            "description": "The Agent returned the bounded MainServerProfilesDto profile projection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainServerProfilesDto"
                },
                "example": {
                  "activeProfileId": "main-server-testing",
                  "activeBinding": "unknown",
                  "detail": "The active Testing profile remains server-owned."
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected a non-canonical host with host_rejected or a non-HTTPS request with https_required; the response uses the safe Agent problem-details contract.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "The request host is not accepted.",
                  "status": 400,
                  "code": "host_rejected",
                  "correlationId": "example-correlation-id"
                }
              }
            }
          },
          "401": {
            "description": "The Windows authentication middleware issued a Negotiate challenge. This framework response is bodyless and is not guaranteed to contain Agent problem details.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Negotiate challenge emitted by the Windows authentication middleware.",
                "schema": {
                  "type": "string",
                  "example": "Negotiate"
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/main-server/state": {
      "get": {
        "tags": [
          "Main Server"
        ],
        "summary": "Read bound Main Server Branch and POS state",
        "description": "Performs only typed, server-owned GET projections after Branch/POS identity and fixed profile binding succeed. Installation acknowledgements, credentials, free URLs, and generic proxy behavior are outside this read boundary.",
        "operationId": "GetMainServerState",
        "responses": {
          "200": {
            "description": "The Agent returned the bounded MainServerStateEvidenceDto state projection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainServerStateEvidenceDto"
                },
                "example": {
                  "profileId": "main-server-testing",
                  "environment": "testing",
                  "binding": "unknown",
                  "outcome": "notAttempted",
                  "detail": "Main Server state was not attempted because binding was not established."
                }
              }
            }
          },
          "400": {
            "description": "The Agent rejected a non-canonical host with host_rejected or a non-HTTPS request with https_required; the response uses the safe Agent problem-details contract.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                },
                "example": {
                  "type": "about:blank",
                  "title": "The request host is not accepted.",
                  "status": 400,
                  "code": "host_rejected",
                  "correlationId": "example-correlation-id"
                }
              }
            }
          },
          "401": {
            "description": "The Windows authentication middleware issued a Negotiate challenge. This framework response is bodyless and is not guaranteed to contain Agent problem details.",
            "headers": {
              "WWW-Authenticate": {
                "description": "Negotiate challenge emitted by the Windows authentication middleware.",
                "schema": {
                  "type": "string",
                  "example": "Negotiate"
                }
              }
            }
          },
          "403": {
            "description": "AuthorizationMiddleware may reject a non-Administrator with a bodyless response. If the exact-origin transport gate rejects the browser origin, the safe problem code is origin_rejected.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentProblemDetailsDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentProblemDetailsDto": {
        "required": [
          "type",
          "title",
          "status"
        ],
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Problem type identifier for the stable Agent error contract."
          },
          "title": {
            "type": "string",
            "description": "Safe human-readable explanation of why the Agent rejected the request."
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "HTTP status code returned by the Agent.",
            "format": "int32"
          },
          "code": {
            "type": [
              "null",
              "string"
            ],
            "description": "Optional stable machine-readable Agent problem code used by clients; it never carries raw identity or credential data."
          },
          "correlationId": {
            "type": [
              "null",
              "string"
            ],
            "description": "Optional request correlation identifier echoed by the Agent for diagnostics."
          }
        },
        "description": "Safe application/problem+json error contract. Code is a stable problem code and correlationId may identify the request; sensitive identity and machine details are not included."
      },
...
      "MainServerBindingStateDto": {
        "enum": [
          "bound",
          "unavailable",
          "mismatch",
          "ambiguous",
          "unknown"
        ],
        "description": "Typed server-owned state classification; unknown and recovery states are preserved explicitly."
      },
      "MainServerEnvironmentDto": {
        "enum": [
          "testing",
          "production"
        ],
        "description": "Server-owned logical label or profile value; it is not a browser-selected path or endpoint."
      },
      "MainServerProfileDto": {
        "required": [
          "profileId",
          "environment",
          "enabled",
          "binding",
          "clientName",
          "allowedReadOperations",
          "detail"
        ],
        "type": "object",
        "properties": {
          "profileId": {
            "type": "string",
            "description": "Server-owned logical label or profile value; it is not a browser-selected path or endpoint."
          },
          "environment": {
            "description": "Server-owned Main Server environment selection; Production remains disabled until explicitly enabled by the server.",
            "$ref": "#/components/schemas/MainServerEnvironmentDto"
          },
          "enabled": {
            "type": "boolean",
            "description": "Boolean outcome or capability flag produced by the server-owned boundary."
          },
          "binding": {
            "description": "Server-owned Branch/POS binding state; the browser cannot provide a host, branch, POS, or endpoint override.",
            "$ref": "#/components/schemas/MainServerBindingStateDto"
          },
          "clientName": {
            "type": "string",
            "description": "Server-owned logical label or profile value; it is not a browser-selected path or endpoint."
          },
          "allowedReadOperations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MainServerReadOperationDto"
            },
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "detail": {
            "type": "string",
            "description": "Safe operator detail without credentials, raw paths, SQL, shell text, private keys, or exception data."
          }
        },
        "description": "Safe projection of one server-owned Main Server profile and its fixed logical read operations."
      },
      "MainServerProfilesDto": {
        "required": [
          "profiles",
          "activeProfileId",
          "activeBinding",
          "detail"
        ],
        "type": "object",
        "properties": {
          "profiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MainServerProfileDto"
            },
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "activeProfileId": {
            "type": "string",
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "activeBinding": {
            "description": "Server-owned Branch/POS binding state for the active profile; the browser cannot provide a host, branch, POS, or endpoint override.",
            "$ref": "#/components/schemas/MainServerBindingStateDto"
          },
          "detail": {
            "type": "string",
            "description": "Safe operator detail without credentials, raw paths, SQL, shell text, private keys, or exception data."
          }
        },
        "description": "Server-owned Testing and future Production Main Server profile projections. URLs, credentials, arbitrary routes, and installation acknowledgements are not returned."
      },
      "MainServerReadOperationDto": {
        "enum": [
          "branchStatus",
          "installedBranch",
          "installedPos"
        ]
      },
      "MainServerReadOutcomeDto": {
        "enum": [
          "notAttempted",
          "succeeded",
          "actionRequired",
          "outcomeUnknown"
        ],
        "description": "Typed server-owned state classification; unknown and recovery states are preserved explicitly."
      },
      "MainServerStateEvidenceDto": {
        "required": [
          "profileId",
          "environment",
          "binding",
          "outcome",
          "branchCode",
          "posNumber",
          "clientName",
          "branchState",
          "posState",
          "detail",
          "checkedAtUtc",
          "correlationId"
        ],
        "type": "object",
        "properties": {
          "profileId": {
            "type": "string",
            "description": "Server-owned logical label or profile value; it is not a browser-selected path or endpoint."
          },
          "environment": {
            "description": "Server-owned Main Server environment associated with this state observation.",
            "$ref": "#/components/schemas/MainServerEnvironmentDto"
          },
          "binding": {
            "description": "Server-owned Branch/POS binding associated with this state observation.",
            "$ref": "#/components/schemas/MainServerBindingStateDto"
          },
          "outcome": {
            "description": "Typed Main Server read outcome; unavailable or ambiguous responses remain explicit.",
            "$ref": "#/components/schemas/MainServerReadOutcomeDto"
          },
          "branchCode": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "posNumber": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "clientName": {
            "type": [
              "null",
              "string"
            ],
            "description": "Server-owned logical label or profile value; it is not a browser-selected path or endpoint."
          },
          "branchState": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "posState": {
            "type": [
              "null",
              "string"
            ],
            "description": "Safe server-owned Slice B field; unrestricted machine details are not exposed."
          },
          "detail": {
            "type": "string",
            "description": "Safe operator detail without credentials, raw paths, SQL, shell text, private keys, or exception data."
          },
          "checkedAtUtc": {
            "type": "string",
            "description": "UTC lifecycle timestamp produced and enforced by the Agent.",
            "format": "date-time"
          },
          "correlationId": {
            "type": "string",
            "description": "Safe request correlation identifier for operator diagnostics."
          }
        },
        "description": "Bounded Branch/POS Main Server state evidence with explicit binding and read outcome truth."
      },
...
    },
    "securitySchemes": {
      "WindowsNegotiate": {
        "type": "http",
        "description": "Windows Integrated Authentication using Negotiate. The browser/OS supplies credentials with a credentialed cross-origin request; clients must rely on the browser's integrated credential flow.",
        "scheme": "negotiate"
      }
    }
  },
...
}